+++ src/datumaro/cli/util/compare.py
@@ -21,14 +21,11 @@ from datumaro.util.image import save_image
 from datumaro.util.import_util import lazy_import
 
 if TYPE_CHECKING:
-    import matplotlib.pyplot as plt
-
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
         import tensorboardX as tb
 else:
     tb = lazy_import("tensorboardX")
-    plt = lazy_import("matplotlib.pyplot")
 
 
 class DistanceCompareVisualizer:
@@ -292,6 +289,8 @@ class DistanceCompareVisualizer:
         self._file_writer.add_image(name, img)
 
     def save_conf_matrix(self, conf_matrix, filename):
+        import matplotlib.pyplot as plt
+
         def _get_class_map(label_categories):
             classes = None
             if label_categories is not None: